+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## 0.0.1
 
 ## 0.0.3
-
 1. Steps Count Stats.
+
+## 0.0.4
+1. Water Stats.
+
+## 0.0.5
+1. Sleep Stats.
+
+## 0.0.6
+1. Mindfulness Stats.
+